BackEnd/game-diffusion-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]
passlib[bcrypt]
torch
opencv-python
motor
google-auth
requests